Fixing bug R-000090: "Internal Error when address updated"
The address' module was trying to reload the address details after the interface was no longer showing.  This
caused the service URL to be invalid which caused a redirect error.  The solution has been added to alter
CnBaseViewFactory::onView() such that the function does not attempt to load view details if the state and
model's subjects do not match.